fixed issue with setting values + fixed array out of bounds problem + change cpu util to int in gui

+++ Cloud-Profiling-Simulation/src/de/hpi_web/cloudSim/profiling/gui/SetValuesAction.java
@@ -16,7 +16,7 @@ public class SetValuesAction implements ActionListener{
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
-		start.setDelay(Integer.parseInt(parent.getTextField_2().getText()));
+		start.setDelay(Double.parseDouble(parent.getTextField_2().getText()));
 		start.setUpperThreshold(Integer.parseInt(parent.getTextField().getText()));
 		start.setLowerThreshold(Integer.parseInt(parent.getTextField_1().getText()));
 	}
